+++ modules/siddhi-core/src/main/java/io/siddhi/core/query/input/stream/join/JoinProcessor.java
@@ -44,6 +44,7 @@ import java.util.List;
  */
 public class JoinProcessor implements Processor {
 
+    // TODO: 2020-09-10  
     private static final Logger log = Logger.getLogger(JoinProcessor.class);
     private boolean trigger;
     private boolean leftJoinProcessor = false;
@@ -76,6 +77,7 @@ public class JoinProcessor implements Processor {
      *
      * @param complexEventChunk event chunk to be processed
      */
+    // TODO: 2020-09-10 catch here 
     @Override
     public void process(ComplexEventChunk complexEventChunk) {
         if (trigger) {
